pyproject.toml
[project]
name = "dream-team-test-task"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "pytest>=9.0.2",
    "python-socketio>=5.16.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
